+++ src/Molecules/Input/FormField/FormField.tsx
@@ -97,16 +97,17 @@ export const FormField: React.FC<Props> = React.forwardRef<HTMLDivElement, Props
     {
       children,
       className,
-      width,
-      label,
-      hideLabel,
-      size,
-      error,
-      success,
       disabled,
+      error,
       extraInfo,
-      onClick,
+      hideLabel,
       innerWrapperRef,
+      label,
+      onClick,
+      required,
+      size,
+      success,
+      width,
     },
     ref,
   ) => (
@@ -120,6 +121,7 @@ export const FormField: React.FC<Props> = React.forwardRef<HTMLDivElement, Props
             color={t => t.color.label}
           >
             {label}
+            {required ? ' *' : ''}
           </HidableTypography>
           <Flexbox item alignSelf="stretch">
             <Typography type="secondary" color={t => t.color.text} as="div">
@@ -138,14 +140,10 @@ export const FormField: React.FC<Props> = React.forwardRef<HTMLDivElement, Props
       <AnimatePresence>
         {hasError(error) ? (
           <BottomWrapper
-            size={size}
             initial={{ y: -10, opacity: 0 }}
             animate={{ y: 0, opacity: 1 }}
             exit={{ y: 0, opacity: 0 }}
             aria-live="polite"
-            // TODO: Check this one @manman
-            // @ts-ignore
-            aria-relevant="additions removals"
           >
             <DensedTypography type="tertiary" color={t => t.color.negative}>
               <VisuallyHidden>Error: </VisuallyHidden>