+++ src/Molecules/Input/index.ts
@@ -1,6 +1,7 @@
 import { Text } from './Text';
 import { Select } from './Select';
+import { Number } from './Number';
 
 // Stitch all inputs so
 // we can use e.g. Input.Text
-export default { Text, Select };
+export default { Text, Number, Select };