+++ src/Molecules/Link/Link.tsx
@@ -23,14 +23,14 @@ const getSharedStyle = (props: ThemedStyledProps<LinkProps, Theme>) => {
   `;
 };
 
-const CleanLink = (props: LinkProps) => {
+const CleanLink = React.forwardRef((props: LinkProps, ref) => {
   return props.external ? (
     // eslint-disable-next-line jsx-a11y/anchor-has-content
-    <a {...R.omit(['fullWidth', 'colorFn', 'color', 'display'], props) as any} />
+    <a ref={ref} {...R.omit(['fullWidth', 'colorFn', 'color', 'display'], props) as any} />
   ) : (
-    <RouterLink {...R.omit(['fullWidth', 'colorFn', 'color', 'display'], props) as any} />
+    <RouterLink ref={ref} {...R.omit(['fullWidth', 'colorFn', 'color', 'display'], props) as any} />
   );
-};
+});
 
 const StyledLink = styled(CleanLink)<LinkProps>`
   ${p => getSharedStyle(p)}
@@ -39,8 +39,9 @@ const StyledLink = styled(CleanLink)<LinkProps>`
 
 const StyledButton = styled(NormalizedElements.Button)<LinkProps>`
   ${p => getSharedStyle(p)}
-
   /* resetting button styles */
+  background: none;
+  cursor: ${p => (p.disabled ? 'not-allowed' : 'pointer')};
   border: none;
   margin: 0;
   padding: 0;
@@ -62,8 +63,18 @@ const StyledButton = styled(NormalizedElements.Button)<LinkProps>`
   font-weight: inherit; /* remove when and if typography is handled inside the component */
 `;
 
-export const Link: LinkComponent = React.forwardRef((props, ref) => {
-  const { to, children, disabled, className, onClick, target, rel, external, as } = props;
+export const Link: LinkComponent = React.forwardRef<any, LinkProps>((props, ref) => {
+  const {
+    to,
+    children,
+    disabled,
+    className,
+    onClick,
+    external,
+    target = external ? '_blank' : undefined,
+    rel = external ? 'noopener noreferrer nofollow' : undefined,
+    as,
+  } = props;
   const destinationProp = external ? { href: to } : { to };
 
   if (isUndefined(to) || disabled) {