+++ src/Molecules/Modal/ModalInner.tsx
@@ -2,63 +2,50 @@ import React, { useEffect, useState } from 'react';
 import { RemoveScroll } from 'react-remove-scroll';
 import FocusLock from 'react-focus-lock';
 import styled from 'styled-components';
+import { motion } from 'framer-motion';
 import { InnerProps, DialogProps } from './Modal.types';
 import NormalizedElements from '../../common/NormalizedElements';
 import { isFunction } from '../../common/utils';
 import { Flexbox, Typography, Icon, Box, useKeyPress } from '../..';
 
-const Backdrop = styled.div`
+const Backdrop = styled(Flexbox)`
+  position: fixed;
+  top: 0;
+  left: 0;
+  height: 100%;
+  width: 100%;
+  z-index: 2;
+
   ${({ theme }) => theme.media.greaterThan(theme.breakpoints.sm)} {
     background-color: ${({ theme }) => theme.color.modalBackdrop};
-    position: fixed;
-    top: 0;
-    left: 0;
-    height: 100%;
-    width: 100%;
-    z-index: 2;
   }
 `;
 
-const Dialog = styled.div<DialogProps>`
+export const Dialog = styled(motion.div)<DialogProps>`
   box-sizing: border-box;
   padding: ${({ theme }) => theme.spacing.unit(5)}px;
   border: 0;
   background: ${({ theme }) => theme.color.card};
-  position: fixed;
-  top: 0;
-  left: 0;
-  z-index: 3;
 
   ${({ theme }) => theme.media.lessThan(theme.breakpoints.sm)} {
     width: 100%;
     height: 100%;
+    transform: none !important; /* disables the appear animation */
   }
 
   ${({ theme }) => theme.media.greaterThan(theme.breakpoints.sm)} {
     width: ${({ theme }) => theme.spacing.unit(120)}px;
-    top: 50%;
-    left: 50%;
     box-shadow: 0 2px 2px 0 ${({ theme }) => theme.color.shadowModal};
-    transform: translate(-50%, 0%);
-    opacity: 0;
-    transition: transform 0.3s ease-in, opacity 0.3s ease-in;
-    ${({ show }) =>
-      show
-        ? `
-      transform: translate(-50%, -50%);
-      opacity: 1;
-    `
-        : ''}
   }
 `;
 
 const Button = styled(NormalizedElements.Button)`
+  display: flex;
   background: none;
   margin-left: ${p => p.theme.spacing.unit(4)}px;
   padding: 0;
   border: 0;
   cursor: pointer;
-  display: flex;
 `;
 
 const RightAlignedFlex = styled(Flexbox)`
@@ -73,6 +60,21 @@ const Content = styled.div`
 export const ModalInner: React.FC<InnerProps> = ({ children, className, title, onClose }) => {
   const [show, setShow] = useState(false);
   const escapePress = useKeyPress('Escape');
+  const animationProps = {
+    initial: {
+      opacity: 0,
+      y: 70,
+    },
+    animate: {
+      opacity: 1,
+      y: 0,
+    },
+    transition: {
+      type: 'spring',
+      damping: 15,
+      stiffness: 200,
+    },
+  };
 
   useEffect(() => {
     setShow(true); // Show is only used for animation
@@ -90,33 +92,34 @@ export const ModalInner: React.FC<InnerProps> = ({ children, className, title, o
     <>
       <FocusLock>
         <RemoveScroll>
-          <Dialog
-            className={className}
-            role="dialog"
-            show={show}
-            {...(title ? { 'aria-label': title } : {})} // TODO: move to aria-labeledby when SSR uid works
-          >
-            <Box mb={2}>
-              <Flexbox container alignItems="baseline">
-                {title && (
-                  <Flexbox item>
-                    <Typography as="h2" type="title2">
-                      {title}
-                    </Typography>
-                  </Flexbox>
-                )}
-                <RightAlignedFlex item>
-                  <Button type="button" onClick={onClose}>
-                    <Icon.Cross size={5} title="Close this dialog" />
-                  </Button>
-                </RightAlignedFlex>
-              </Flexbox>
-            </Box>
-            <Content>{children}</Content>
-          </Dialog>
+          <Backdrop className={className} container alignItems="center" justifyContent="center">
+            <Dialog
+              role="dialog"
+              show={show}
+              {...(title ? { 'aria-label': title } : {})} // TODO: move to aria-labeledby when SSR uid works
+              {...animationProps}
+            >
+              <Box mb={2}>
+                <Flexbox container alignItems="baseline">
+                  {title && (
+                    <Flexbox item>
+                      <Typography as="h2" type="title2">
+                        {title}
+                      </Typography>
+                    </Flexbox>
+                  )}
+                  <RightAlignedFlex item>
+                    <Button type="button" onClick={onClose}>
+                      <Icon.Cross size={5} title="Close this dialog" />
+                    </Button>
+                  </RightAlignedFlex>
+                </Flexbox>
+              </Box>
+              <Content>{children}</Content>
+            </Dialog>
+          </Backdrop>
         </RemoveScroll>
       </FocusLock>
-      <Backdrop />
     </>
   );
 };