+++ src/common/utils.ts
@@ -41,6 +41,7 @@ export const deprecate = (message: string) => <T extends {} | Function>(target: 
 export const isUndefined = (x: any): x is undefined => typeof x === 'undefined';
 export const isElement = (x: any): x is React.ReactNode => React.isValidElement(x);
 export const isNumber = (x: any): x is number => typeof x === 'number';
+export const isString = (x: any): x is string => typeof x === 'string';
 export const isBoolean = (x: any): x is boolean => typeof x === 'boolean';
 export const isFunction = (x: any): x is Function => typeof x === 'function';
 export const isHTMLElement = (x: any): x is HTMLElement => x instanceof HTMLElement;