+++ src/index.ts
@@ -6,6 +6,7 @@ import CssGrid from './Atoms/CssGrid';
 import DateTime from './Atoms/DateTime';
 import Flag from './Atoms/Flag';
 import Flexbox from './Atoms/Flexbox';
+import FormLabel from './Atoms/FormLabel';
 import Icon from './Atoms/Icon';
 import List from './Atoms/List';
 import ListItem from './Atoms/ListItem';
@@ -44,7 +45,7 @@ import CardWithTabs from './Organisms/CardWithTabs';
 import Development from './Organisms/Development';
 
 /** Hooks */
-import useKeyPress from './common/Hooks/useKeyPress';
+import { useKeyPress } from './common/Hooks';
 
 import theme, { createTheme } from './theme';
 
@@ -64,6 +65,7 @@ export {
   FeedbackBanner,
   Flag,
   Flexbox,
+  FormLabel,
   Icon,
   Input,
   LabeledValue,